requirements.txt
Pillow==5.2.0
requests==2.18.4
simplejson==3.11.1
